Fix default xc-tools.iso behavior
The new/corrected behavior:
-During VM creation, if the xc-tools.iso is present in /storage/isos and
autostart is not set to "iso" then xc-tools.iso is set to be the default cd.
Otherwise, it is set to "None" ("").

-On the VM Details screen Hardware tab, if xc-tools.iso is present then
it will appear in the drop down as "OpenXT Guest Tools".  Any other ISO
will appear as it's file name.

Note: the list of available isos is currently updated only on the initial UI load.

OXT-227

Signed-off-by: Andrew 'Doc' Docherty <doc@coveycs.com>