VM Disks are no longer deleteable when modify-vm-settings is false
When the MODEL_DISKS_CHANGED topic is published, bind is called
and the repeater displaying the disks is reloaded
but it doesn't check to see if the disks are allowed to be
deleted or not, so the button is re-enabled.

Solution was to add that check in after the bind.

OXT-98

Signed-off-by: Ian Miller <ian@coveycs.com>