3D Graphics PCI Pass Through Prompt
Add a dialog that prompts the user so that they are aware of any
other PCI devices on the 3d card's bus that will be passed through
to the guest.

Expected behavior:

In the VM Details dialog, on hardware tab, the user makes a selection
in the "3D Graphics (experimental)" drop down.  A dialog should appear
if:

- There are other devices on the gpu bus that will be passed through
  AND
- The use makes a selection besides disabled (either Enabled or the
specific card on multi-gpu boxes)

A dialog will appear warning a user about the other devices being
passed through, along with a list of those devices.

The user may click "OK" to dismiss the dialog, or "Cancel" which
sets the drop down to the last saved setting.  The top right "X" works
the same as "Cancel".

OXT-281

Signed-off-by: Andrew 'Doc' Docherty <doc@coveycs.com>